backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=2.0.0

# LangChain Core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# LlamaIndex for RAG
llama-index-core>=0.11.0
llama-index-readers-file>=0.2.0
llama-index-embeddings-huggingface>=0.4.0
llama-index-retrievers-bm25>=0.4.0
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
httpx>=0.26.0
aiofiles>=23.2.1

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0

# Crypto for platform encryption
pycryptodome>=3.20.0

# Feishu SDK for long connection
lark-oapi>=1.0.0
# Data Analysis & Visualization
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
tabulate>=0.9.0
# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
nest_asyncio
